backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "^3.13.1"
Django = "^4.0.6"
feedparser = "^6.0.10"
django-filter = "^22.1"
django-cors-headers = "^3.13.0"
tweepy = "^4.10.0"
psycopg2-binary = "^2.9.3"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
